+++ lib/extended_email_reply_parser/parsers/github.rb
@@ -1,0 +1,9 @@
+module ExtendedEmailReplyParser
+  class Parsers::Github < Parsers::Base
+
+    def parse
+      EmailReplyParser.parse_reply text
+    end
+
+  end
+end